# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "V.SMART API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/currencies": {
      "get": {
        "tags": [
          "Currency"
        ],
        "summary": "M2-B02 — the reference implementation of the paged list contract (ADR-002 §2 and its\r\nM2-B02 addendum). Every future list endpoint copies this shape: one\r\n`[FromQuery]` typed query record in, one V.SMART.Api.Contracts.PagedResult`1 out, the\r\nservice's filter dictionary produced by V.SMART.Api.Contracts.FilterDictionaryAdapter and\r\nnever seen on the wire.\r\n            \r\n\r\nThe controller-local paged-response record it replaces is deleted: 60–80copies of it would become 60–80 interfaces in the generated TypeScript client. TheJSON is unchanged — the four property names were already ADR-002's.\r\n<b>Behaviour change to note:</b> the default `pageSize` is now thecontract-wide 20, where this endpoint alone previously defaulted to 10. Callers thatsend `pageSize` explicitly are unaffected.\r\nValidation is declarative on V.SMART.Api.Contracts.CurrencyQuery, so an invalid`pageNumber`/`pageSize`/`sort`/date range is rejected by M2-A06's`InvalidModelStateResponseFactory` before this method runs. The explicit`ModelState` guard below is the belt-and-braces path for a caller that invokes theaction directly (unit tests), and matches `Create`/`Update`.",
        "operationId": "getCurrencies",
        "parameters": [
          {
            "name": "currName",
            "in": "query",
            "description": "Case-insensitive contains match on the currency name. Wire name `currName`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdBy",
            "in": "query",
            "description": "Case-insensitive contains match on the creating user. Wire name `createdBy`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "description": "Inclusive lower bound on `CreatedDate`, applied at the start of the day.\r\n            Wire name `fromDate`.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "description": "Inclusive upper bound on `CreatedDate`, applied at the end of the day.\r\n            Wire name `toDate`.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "1-based page index. Default 1. Wire name `pageNumber`.",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Rows per page. Default 20, maximum 100. Wire name `pageSize`.",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields, `-` prefix for descending — e.g.\r\n`-createdDate,currName`. Omit it to keep the resource's existing default ordering.\r\nValidated against V.SMART.Api.Contracts.PagedQuery.SortableFields; an unknown field is a 400 that lists\r\nthe permitted values. Wire name `sort`.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVMPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVMPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVMPagedResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Currency"
        ],
        "summary": "Creates a currency. A duplicate name is the service's refusal, surfaced as 409 with the\r\nservice's message verbatim in `title` (ADR-002 §4); a malformed body is 400 with the\r\nper-field `errors` dictionary.",
        "operationId": "createCurrency",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyVM"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyVM"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CurrencyVM"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyVM"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reference/gst-rates": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "The two GST ladders, paired by index. Read from `CommonConstants`, never retyped.",
        "description": "This action is not tenant-dependent — the ladders are compile-time constants — but it\r\nshares the group's tenant-keyed policy anyway. Carving out a per-endpoint exemption\r\nwould buy a marginal cache-entry saving in exchange for two policies to reason about,\r\nand the failure mode of getting that wrong is a cross-tenant leak. Uniform is safer.",
        "operationId": "getGstRates",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GstRatesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GstRatesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GstRatesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reference/uoms": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "Units of measure. `UnitCode` is the key; this table has no integer id.",
        "operationId": "getUoms",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UomDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UomDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UomDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reference/states": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "States.",
        "operationId": "getStates",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StateDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StateDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StateDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reference/terms": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "Active terms and conditions. The active filter is the domain's — this action calls\r\n`GetAllActiveTermsAsync` and does not re-filter, so the API cannot disagree with\r\nthe Blazor screens about what \"active\" means.",
        "operationId": "getTerms",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TermsDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TermsDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TermsDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reference/screens": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "The screen catalogue — the permission <b>vocabulary</b>, not the caller's rights.",
        "description": "\r\n<b>The protection decision, recorded here because it is a security decision.</b>            This returns every screen name in the tenant, which is a map of the application's            surface. It is <i>not</i> sensitive per-user data: the same list is identical for            every user in the tenant, and it discloses nothing about who may do what. The            caller's own rights are a different endpoint (`GET /api/v1/me`, M2-A07) and are            deliberately not merged into this one — merging them would make this response            caller-dependent and force the cache key to include the user, multiplying the entry            count by the user base for no benefit.\r\n\r\nIt therefore requires authentication, is tenant-keyed, and is cached like the            rest of the group. It projects to V.SMART.Api.Contracts.ScreenDto specifically to drop the            `UserRights` navigation — returning the entity would put the tenant's entire            permission matrix on the wire behind a dropdown feed.",
        "operationId": "getScreens",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScreenDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScreenDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScreenDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reference/currencies": {
      "get": {
        "tags": [
          "Reference"
        ],
        "summary": "Currencies, as reference data for dropdowns.",
        "description": "<b>Distinct from `CurrencyController`, deliberately.</b> That controller is the\r\n            CRUD surface for the Currency master — paged, sorted, filtered, writable\r\n            (`/api/v1/currencies`). This is a flat, cached, read-only list for populating a\r\n            selector (`/api/v1/reference/currencies`). Different lifetime, different shape,\r\n            different cacheability; the route group and these summaries are what keep them\r\n            distinguishable in Swagger. It excludes the `CurrencyRates` navigation — the\r\n            daily rate feed changes on a different clock and is what makes the entity unsafe to\r\n            cache.",
        "operationId": "getReferenceCurrencies",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrencyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrencyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrencyDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CurrencyVM": {
        "required": [
          "currName",
          "currSub",
          "symbol"
        ],
        "type": "object",
        "properties": {
          "currId": {
            "type": "integer",
            "format": "int32"
          },
          "currName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "currSub": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "symbol": {
            "minLength": 1,
            "pattern": "[$€₹¥£₩₿₽]",
            "type": "string"
          },
          "isSystemDefined": {
            "type": "boolean"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CurrencyVMPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurrencyVM"
            },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "The single paged-list response body for every list endpoint in this API (M2-B02).\r\n            \r\n\r\nThe property names are ADR-002 §2's `{ items, totalCount, pageNumber, pageSize }`verbatim; with ASP.NET Core's default camel-case JSON policy they serialise to exactlythose four names. This is <b>one generic type</b>, deliberately replacing thecontroller-local paged-response record pattern (the one `CurrencyController` used to declare): 60–80 structurally identicalresponse records would become 60–80 distinct interfaces in the generated TypeScriptclient (M2-B10).\r\n<b>Why there is no `totalPages`.</b> ADR-002 §2 names four properties and thegenerated client is frozen at M2-B03, so adding a fifth here is a contract decision, not aconvenience. It is derivable on the client from `totalCount` and `pageSize`.Note that an unrelated, unused `V.SMART.Shared.ViewModels.PagedResult<T>`(`RejectionMasterVM.cs:33-40`) does carry `TotalPages` — it has zero referencesand is not this type. Do not confuse the two; if a file needs both namespaces it mustalias one.\r\n`TotalCount` is the count of the <b>filtered, unpaged</b> query — see`CurrencyService.cs:76`, where `CountAsync()` runs after the filters and before`Skip`/`Take`. Every service in `BusinessLayer/` follows that convention(INV-002); a list endpoint that reports anything else is a defect."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}